Make generation of Add/UpdateX query a common function
